+++ src/main/java/persistence/model/EntityFields.java
@@ -9,7 +9,6 @@ import java.util.List;
 public class EntityFields {
     private EntityField idField;
     private final List<EntityField> fields = new ArrayList<>();
-    private final List<EntityJoinEntityField> joinFields = new ArrayList<>();
 
     public void addField(final EntityField field) {
         if (field.getField().isAnnotationPresent(Id.class)) {
@@ -22,15 +21,7 @@ public class EntityFields {
         return this.idField;
     }
 
-    public void addJoinField(final EntityJoinEntityField joinEntityField) {
-        this.joinFields.add(joinEntityField);
-    }
-
     public List<EntityField> getFields() {
         return Collections.unmodifiableList(this.fields);
-    }
-
-    public List<EntityJoinEntityField> getJoinFields() {
-        return Collections.unmodifiableList(this.joinFields);
     }
 }
